In order to remove use_sparse_data and use_sparse_alg cruft from compute_Hastings_correction in preparation for new compressed data structure, handle the case where using compression and a single thread in which case the Hastings correction uses fast_sparse_array and needs vector lookups therein.

+++ StochasticBlockPartition/code/python/partition_baseline_main.py
@@ -421,7 +421,7 @@ def propose_node_movement(current_node, partition, out_neighbors, in_neighbors, 
 
 def coo_to_flat(x, size):
     x_i, x_v = x
-    f = np.zeros(size)
+    f = np.zeros(size, dtype=int)
     f[x_i] = x_v
     return f
